autotrader/pyproject.toml
[tool.poetry]
name = "autotrader"
version = "0.1.0"
description = "Multi-model autonomous crypto trading stack"
authors = ["Your Team <team@example.com>"]
packages = [{include = "libs"}, {include = "services"}]

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.7.1"
httpx = "^0.27"
openai = "^1.23"
ujson = "^5.10.0"
fastapi = "^0.111"
uvicorn = {extras=["standard"], version="^0.29"}
kafka-python = "^2.0"
qdrant-client = "^1.8"
sentence-transformers = "^2.7"
faiss-cpu = "^1.8"
websocket-client = "^1.8"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2"
ruff = "^0.4"
mypy = "^1.10"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
